+++ bin/post_processing.py
@@ -96,14 +96,13 @@ def process_results(db_dir,out_dir, prefix, gene_predictor):
     merged_df["annot"] = merged_df["annot"].replace(nan, 'hypothetical protein', regex=True)
     merged_df["category"] = merged_df["category"].replace(nan, 'unknown function', regex=True)
     merged_df["color"] = merged_df["color"].replace(nan, 'None', regex=True)
+
     # process vfdb results
-    merged_df = process_vfdb_results(out_dir, merged_df)
+    (merged_df, vfdb_results) = process_vfdb_results(out_dir, merged_df)
     # process CARD results
-    merged_df = process_card_results(out_dir, merged_df, db_dir)
-    # write to final output
-    merged_df.to_csv( os.path.join(out_dir, prefix + "_final_merged_output.tsv"), sep="\t", index=False)
-    
-    return merged_df
+    (merged_df, card_results) = process_card_results(out_dir, merged_df, db_dir)
+
+    return (merged_df,vfdb_results, card_results)
 
 def get_contig_name_lengths(fasta_input):
     """
@@ -275,30 +274,43 @@ def create_gff(cds_mmseqs_df, length_df, fasta_input, out_dir, prefix, locustag,
     :param prefix: output prefix
     :param locustag: whether or not to create a random locustag - will be Random is so. Otherwise it is parsed
     :tmrna_flag boolean whether there are tmRNAs or not
-    :return: locustag for the creation of the .tbl file
+    :return: locustag for the creation of the .tbl file, locus_df as df with consistent locus_tags
     """
-    #########
-    # rearrange start and stop so that start is always less than stop for gff
-    #########
-    cols = ["start","stop"]
-    #indices where start is greater than stop
-    ixs = cds_mmseqs_df['frame'] == '-'
-    # Where ixs is True, values are swapped
-    cds_mmseqs_df.loc[ixs,cols] = cds_mmseqs_df.loc[ixs, cols].reindex(columns=cols[::-1]).values
+
+    # create locus tag and get temp df
+
 
     # locustag creation
     if locustag == "Random":
         # locus tag header 8 random letters
         locustag = ''.join(random.choice(string.ascii_uppercase) for _ in range(8))
     
+    ############ locus tag #########
+    # write df for locus tag parsing
+    locus_df = cds_mmseqs_df
+    locus_df['locus_tag'] = locustag + "_CDS_" + cds_mmseqs_df.index.astype(str)
+    #################################
+
+    #########
+    # rearrange start and stop so that start is always less than stop for gff
+    #########
+
+    cols = ["start","stop"]
+    #indices where start is greater than stop
+    ixs = cds_mmseqs_df['frame'] == '-'
+    # Where ixs is True, values are swapped
+    cds_mmseqs_df.loc[ixs,cols] = cds_mmseqs_df.loc[ixs, cols].reindex(columns=cols[::-1]).values
+
     # set phase to be 0
     cds_mmseqs_df['phase'] = 0
     # create attributes
-    cds_mmseqs_df['attributes'] = "ID=" + locustag + "_CDS_" + cds_mmseqs_df.index.astype(str)  + ";" + "phrog=" + cds_mmseqs_df["phrog"].astype(str) + ";" + "top_hit=" + cds_mmseqs_df["top_hit"].astype(str) + ";" + "locus_tag=" + locustag + "_" + cds_mmseqs_df.index.astype(str) + ";" + "function=" + cds_mmseqs_df["category"].astype(str) + ";"  + "product=" + cds_mmseqs_df["annot"].astype(str)
+    cds_mmseqs_df['attributes'] = "ID=" + locus_df['locus_tag'].astype(str)  + ";" + "phrog=" + cds_mmseqs_df["phrog"].astype(str) + ";" + "top_hit=" + cds_mmseqs_df["top_hit"].astype(str) + ";" + "locus_tag=" + locus_df['locus_tag'].astype(str) + ";" + "function=" + cds_mmseqs_df["category"].astype(str) + ";"  + "product=" + cds_mmseqs_df["annot"].astype(str)
     # adds VFDB
     cds_mmseqs_df.loc[cds_mmseqs_df['vfdb_short_name'] != "None", 'attributes'] = cds_mmseqs_df['attributes'].astype(str)  + ";"  + "vfdb_short_name=" + cds_mmseqs_df['vfdb_short_name'].astype(str) + ";"  + "vfdb_description="  +  cds_mmseqs_df['vfdb_description'].astype(str) + ";" + "vfdb_species=" + cds_mmseqs_df['vfdb_species'].astype(str)
     # adds CARD
     cds_mmseqs_df.loc[cds_mmseqs_df['CARD_short_name'] != "None", 'attributes'] = cds_mmseqs_df['attributes'].astype(str)  + ";"  + "CARD_short_name=" + cds_mmseqs_df['CARD_short_name'].astype(str) + ";"  + "AMR_Gene_Family="  +  cds_mmseqs_df['AMR_Gene_Family'].astype(str) + ";" + "CARD_species=" + cds_mmseqs_df['CARD_species'].astype(str)
+
+    # save back 
 
     # get gff dataframe in correct order 
     gff_df = cds_mmseqs_df[["contig", "Method", "Region", "start", "stop", "score", "frame", "phase", "attributes"]]
@@ -307,10 +319,9 @@ def create_gff(cds_mmseqs_df, length_df, fasta_input, out_dir, prefix, locustag,
     gff_df["start"] = gff_df["start"].astype('int')
     gff_df["stop"] = gff_df["stop"].astype('int')
 
-    # write to tmp gff
-    with open(os.path.join(out_dir, "phrokka_tmp.gff"), 'w') as f:
+    with open(os.path.join(out_dir, "phrokka_tmp.gff"), 'a') as f:
         gff_df.to_csv(f, sep="\t", index=False, header=False)
-      
+
     ### trnas
     # check if no trnas
     col_list = ["contig", "Method", "Region", "start", "stop", "score", "frame", "phase", "attributes"]
@@ -384,7 +395,169 @@ def create_gff(cds_mmseqs_df, length_df, fasta_input, out_dir, prefix, locustag,
         fasta_sequences = SeqIO.parse(open(fasta_input),'fasta')
         SeqIO.write(fasta_sequences, f, "fasta")
 
-    return locustag
+    return (locustag, locus_df)
+
+
+def update_fasta_headers(locus_df, out_dir, gene_predictor ):
+    """
+    Updates the fasta output headers to have a consistent locus tag & gene description for downstrea use
+    :param locus_df a pandas df as output from create_gff()
+    :param out_dir: output directory path
+    :gene_predictor: string 'phanotate' or 'prodigal' with the gene predictor used
+    """
+
+    #define outputs
+    fasta_input_nts_tmp = gene_predictor + "_out_tmp.fasta"
+    fasta_input_aas_tmp = gene_predictor + "_aas_tmp.fasta"
+    fasta_output_nts_gd = gene_predictor + ".ffn"
+    fasta_output_aas_gd = gene_predictor + ".faa"
+
+    # nucleotides
+
+    with open(os.path.join(out_dir, fasta_output_nts_gd), 'w') as nt_fa:   
+        i = 0 
+        for dna_record in SeqIO.parse(os.path.join(out_dir, fasta_input_nts_tmp), 'fasta'): 
+            dna_record.id = str(locus_df['locus_tag'].iloc[i]) 
+            dna_record.description = str(locus_df['annot'].iloc[i])
+            SeqIO.write(dna_record, nt_fa, 'fasta')
+            i += 1
+
+    # amino acids
+
+    with open(os.path.join(out_dir, fasta_output_aas_gd), 'w') as aa_fa:   
+        i = 0 
+        for dna_record in SeqIO.parse(os.path.join(out_dir, fasta_input_aas_tmp), 'fasta'): 
+            dna_record.id = str(locus_df['locus_tag'].iloc[i]) 
+            dna_record.description = str(locus_df['annot'].iloc[i])
+            SeqIO.write(dna_record, aa_fa, 'fasta')
+            i += 1
+
+
+def extract_terl(locus_df, out_dir, gene_predictor, logger ):
+    """
+    Updates the fasta output headers to have a consistent locus tag & gene description for downstrea use
+    :param locus_df a pandas df as output from create_gff()
+    :param out_dir: output directory path
+    :gene_predictor: string 'phanotate' or 'prodigal' with the gene predictor used
+    :logger: logger
+    """
+
+    #phanotate
+    fasta_input_nts_tmp = gene_predictor + "_out_tmp.fasta"
+    fasta_input_aas_tmp = gene_predictor + "_aas_tmp.fasta"
+
+    # nucleotide
+
+    with open(os.path.join(out_dir, "terL.ffn"), 'w') as aa_fa:   
+        # i loops over the rows of the dataframe
+        # j counts the number of terLs
+        i = 0 
+        j = 0
+        for dna_record in SeqIO.parse(os.path.join(out_dir, fasta_input_nts_tmp), 'fasta'): 
+            dna_record.id = str(locus_df['locus_tag'].iloc[i]) 
+            dna_record.description = str(locus_df['annot'].iloc[i])
+            if locus_df['annot'].iloc[i] == "terminase large subunit":
+                SeqIO.write(dna_record, aa_fa, 'fasta')
+                # report terL found the first time
+                if j < 1:
+                    print('Terminase large subunit found.')
+                    logger.info("Terminase large subunit found.")
+                    # report multiple found the second time
+                if j == 1:
+                    print('More than one CDS annotated as terminase large subunit found. \nSaving all.') 
+                    logger.info('More than one CDS annotated as terminase large subunit found. \nSaving all.') 
+                j += 1
+            i += 1
+
+
+    # amino acid no need to print
+
+    with open(os.path.join(out_dir, "terL.faa"), 'w') as aa_fa:   
+        # i loops over the rows of the dataframe
+        i = 0 
+        for dna_record in SeqIO.parse(os.path.join(out_dir, fasta_input_aas_tmp), 'fasta'): 
+            dna_record.id = str(locus_df['locus_tag'].iloc[i]) 
+            dna_record.description = str(locus_df['annot'].iloc[i])
+            if locus_df['annot'].iloc[i] == "terminase large subunit":
+                SeqIO.write(dna_record, aa_fa, 'fasta')
+            i += 1
+
+
+def update_final_output(cds_mmseqs_merge_df, vfdb_results, card_results, locus_df, prefix, out_dir ):
+    """
+    Updates the fasta output headers to have a consistent locus tag & gene description for downstrea use
+    :param cds_mmseqs_merge_df: a pandas df as output from process_results()
+    :param locus_df: a pandas df as output from create_gff()
+    :param out_dir: output directory path
+    :param prefix: output prefix
+    :return: 
+    """
+
+    # needed for vfdb card matching later
+    genes_for_vfdb_card = cds_mmseqs_merge_df['gene']
+
+    # return back the cds_mmseqs_merge_df but with the locus tag instead of gene
+    # rename gene with locus_tag
+    locus_tag_series = locus_df['locus_tag']
+    cds_mmseqs_merge_df['gene'] = locus_tag_series
+
+    #########
+    # rearrange start and stop for neg strang
+    #########
+
+    st_cols = ["start","stop"]
+    #indices where start is greater than stop
+    ixs = cds_mmseqs_merge_df['frame'] == '-'
+    # Where ixs is True, values are swapped
+    cds_mmseqs_merge_df.loc[ixs,st_cols] = cds_mmseqs_merge_df.loc[ixs, st_cols].reindex(columns=st_cols[::-1]).values
+
+
+
+    # get a list of columns
+    cols = list(cds_mmseqs_merge_df)
+    # move the column to head of list using index, pop and insert
+    cols.insert(0, cols.pop(cols.index('gene')))
+    cds_mmseqs_merge_df = cds_mmseqs_merge_df.loc[:, cols]
+    # drop last 2 cols 
+    cds_mmseqs_merge_df = cds_mmseqs_merge_df.drop(columns=['phase', 'attributes', 'locus_tag'])
+
+    # write output
+    final_output_path = os.path.join(out_dir, prefix + "_cds_final_merged_output.tsv")
+    cds_mmseqs_merge_df.to_csv( final_output_path, sep="\t", index=False)
+
+    ######################################
+    ##### update vfdb with locus tag #####
+    ###### merge locus into the vfdb ####
+    locus_df['gene'] = genes_for_vfdb_card
+    vfdb_results = vfdb_results.merge(locus_df, how='left', on='gene')
+    # get a list of columns
+    cols = list(vfdb_results)
+    # move the column to head of list using index, pop and insert
+    cols.insert(0, cols.pop(cols.index('locus_tag')))
+    vfdb_results = vfdb_results.loc[:, cols]
+    # keep only desired columns  and save
+    vfdb_results = vfdb_results[['locus_tag', 'vfdb_hit_x', 'vfdb_alnScore_x', 'vfdb_seqIdentity_x', 'start', 'stop', 'frame']]
+    vfdb_results.columns = ['gene', 'vfdb_hit', 'vfdb_alnScore', 'vfdb_seqIdentity', 'start', 'stop', 'frame']
+    vfdb_results = vfdb_results.sort_values(by=['start'])
+    vfdb_results.to_csv(os.path.join(out_dir, "top_hits_vfdb.tsv"), sep="\t", index=False)
+
+    ######################################
+    ##### update card with locus tag #####
+    ###### merge locus into the card ####
+    card_results = card_results.merge(locus_df, how='left', on='gene')
+    # get a list of columns
+    cols = list(card_results)
+    # move the column to head of list using index, pop and insert
+    cols.insert(0, cols.pop(cols.index('locus_tag')))
+    card_results = card_results.loc[:, cols]
+
+    # keep only desired columns   sand save
+    card_results = card_results[['locus_tag', 'CARD_hit_x', 'CARD_alnScore_x', 'CARD_seqIdentity_x', 'start', 'stop', 'frame']]
+    card_results.columns = ['gene', 'card_hit', 'card_alnScore', 'card_seqIdentity', 'start', 'stop', 'frame']
+    card_results = card_results.sort_values(by=['start'])
+    card_results.to_csv(os.path.join(out_dir, "top_hits_card.tsv"), sep="\t", index=False)
+
+
 
 def create_tbl(cds_mmseqs_df, length_df, out_dir, prefix, gene_predictor, tmrna_flag, locustag):
     """
@@ -685,7 +858,6 @@ def process_vfdb_results(out_dir, merged_df):
         tophits.append([tmp_df.vfdb_hit, tmp_df.gene, tmp_df.vfdb_alnScore, tmp_df.vfdb_seqIdentity, tmp_df.vfdb_eVal])
 
     tophits_df = pd.DataFrame(tophits, columns=['vfdb_hit', 'gene', 'vfdb_alnScore', 'vfdb_seqIdentity', 'vfdb_eVal'])
-    tophits_df.to_csv(os.path.join(out_dir, "top_hits_vfdb.tsv"), sep="\t", index=False)
 
     # left join vfdb to merged_df
     tophits_df['gene']=tophits_df['gene'].astype(str)
@@ -722,7 +894,7 @@ def process_vfdb_results(out_dir, merged_df):
         merged_df["vfdb_short_name"] = 'None'
         merged_df["vfdb_description"] = 'None'
         merged_df["vfdb_species"] = 'None'
-    return merged_df
+    return (merged_df, tophits_df)
 
 
     
@@ -750,8 +922,7 @@ def process_card_results(out_dir, merged_df, db_dir):
         tophits.append([tmp_df.CARD_hit, tmp_df.gene, tmp_df.CARD_alnScore, tmp_df.CARD_seqIdentity, tmp_df.CARD_eVal])
 
     tophits_df = pd.DataFrame(tophits, columns=['CARD_hit', 'gene', 'CARD_alnScore', 'CARD_seqIdentity', 'CARD_eVal'])
-    tophits_df.to_csv(os.path.join(out_dir, "top_hits_card.tsv"), sep="\t", index=False)
-
+    
     # left join tophits_df to merged_df
     tophits_df['gene']=tophits_df['gene'].astype(str)
 
@@ -797,6 +968,6 @@ def process_card_results(out_dir, merged_df, db_dir):
         merged_df["Drug_Class"] = 'None'
         merged_df["Resistance_Mechanism"] = 'None'
 
-    return merged_df
+    return (merged_df, tophits_df)