+++ bin/processes.py
@@ -114,14 +114,16 @@ def translate_fastas(out_dir, gene_predictor):
         clean_df = tidy_phanotate_output(out_dir)
         fasta_input_tmp = "phanotate_out_tmp.fasta"
         fasta_output_aas_tmp = "phanotate_aas_tmp.fasta"
-        fasta_output_aas_gd = "phanotate_aas.fasta"
-        fasta_output_nts_gd = "phanotate_nts.fasta"
+        fasta_output_aas_gd = "phanotate.faa"
+        fasta_output_nts_gd = "phanotate.ffn"
     if gene_predictor == "prodigal":
         clean_df = tidy_prodigal_output(out_dir)
         fasta_input_tmp = "prodigal_out_tmp.fasta"
         fasta_output_aas_tmp = "prodigal_aas_tmp.fasta"
-        fasta_output_aas_gd = "prodigal_aas.fasta"
-        fasta_output_nts_gd = "prodigal_nts.fasta"
+        fasta_output_aas_gd = "prodigal.faa"
+        fasta_output_nts_gd = "prodigal.ffn"
+    
+    # translate for temporary AA output
     with open(os.path.join(out_dir, fasta_output_aas_tmp), 'w') as aa_fa:
         i = 0 
         for dna_record in SeqIO.parse(os.path.join(out_dir, fasta_input_tmp), 'fasta'): 
@@ -130,23 +132,7 @@ def translate_fastas(out_dir, gene_predictor):
             aa_record = SeqRecord(dna_record.seq.translate(to_stop=True), id=dna_header, description = dna_description )
             SeqIO.write(aa_record, aa_fa, 'fasta')
             i += 1
-    with open(os.path.join(out_dir, fasta_output_aas_gd), 'w') as aa_fa:
-        i = 0 
-        for dna_record in SeqIO.parse(os.path.join(out_dir, fasta_input_tmp), 'fasta'): 
-            dna_header = str(clean_df['contig'].iloc[i]) + "_" + str(i) 
-            dna_description = str(clean_df['start'].iloc[i]) + "_" + str(clean_df['stop'].iloc[i])
-            aa_record = SeqRecord(dna_record.seq.translate(to_stop=True), id=dna_header, description = dna_description )
-            SeqIO.write(aa_record, aa_fa, 'fasta')
-            i += 1
-    with open(os.path.join(out_dir, fasta_output_nts_gd), 'w') as aa_fa:
-        i = 0 
-        for dna_record in SeqIO.parse(os.path.join(out_dir, fasta_input_tmp), 'fasta'): 
-            dna_header = str(clean_df['contig'].iloc[i]) + "_" + str(i) 
-            dna_description = str(clean_df['start'].iloc[i]) + "_" + str(clean_df['stop'].iloc[i])
-            aa_record = SeqRecord(dna_record.seq, id=dna_header, description = dna_description )
-            SeqIO.write(aa_record, aa_fa, 'fasta')
-            i += 1
-    
+
 
 def run_trna_scan(filepath_in, out_dir, logger):
     """